pyproject.toml
[tool.poetry]
name = "genshin-alerts"
version = "0.1.0"
description = "Genshin Impact Alert Bot"
packages = [
  { from = "src", include = "genshin_alerts/*" }
]
authors = ["David Lam <david.lam@lamdav.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
click = "^7.1.2"
colorama = "^0.4.4"
"discord.py" = "^1.5.1"
httpx = "^0.16.1"
pydantic = "^1.7.3"
PyYAML = "^5.3.1"
structlog = "^20.2.0"
validators = "^0.18.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
bump2version = "^1.0.1"

[tool.poetry.scripts]
genshin-alerts = "genshin_alerts.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
